+++ CHANGELOG.md
@@ -1,4 +1,8 @@
-Version 0.1.20201221 
+Version 0.1.20201222
+- Fixed some ugly bug in the searchbinary function
+- Added to stdout the vanitykeys found with -v option
+
+Version 0.1.20201221
 - Fixed search by xpoint.
 - Added -e option to skip the sort process whe the file is already sorted.
 - Fixed debugcount when upto N is less than debugcount.